moving around of components
