+++ components.d.ts
@@ -74,7 +74,7 @@ declare module 'vue' {
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
     SelectedMissionsCard: typeof import('./src/components/SolarSystemView/MissionsList/MissionsCard/SelectedMissionsCard.vue')['default']
-    SnackBar: typeof import('./src/components/SnackBar.vue')['default']
+    SnackBar: typeof import('./src/components/Reused/SnackBar.vue')['default']
     SolarSystem: typeof import('./src/components/Reused/SolarSystem.vue')['default']
     SolarSystemInfo: typeof import('./src/components/SolarSystemView/SolarSystemInfo.vue')['default']
     SolarSystemInfoPanel: typeof import('./src/components/GalaxyView/SolarSystemInfoPanel.vue')['default']
@@ -125,6 +125,7 @@ declare module 'vue' {
     VDialog: typeof import('vuetify/lib')['VDialog']
     VDivider: typeof import('vuetify/lib')['VDivider']
     VerifyEmail: typeof import('./src/components/Landing/VerifyEmail.vue')['default']
+    VersionStatus: typeof import('./src/components/GalaxyList/VersionStatus.vue')['default']
     VExpandTransition: typeof import('vuetify/lib')['VExpandTransition']
     VExpansionPanel: typeof import('vuetify/lib')['VExpansionPanel']
     VExpansionPanelContent: typeof import('vuetify/lib')['VExpansionPanelContent']
