--- src/components/Alpha.vue
@@ -1,27 +1,0 @@
-<template>
-  <div class="alpha">
-    <p class="alpha-text">Aplha Version</p>
-  </div>
-</template>
-
-<script>
-export default {
-  name: "Alpha",
-  data() {
-    return {};
-  },
-  methods: {},
-};
-</script>
-
-<style lang="scss" scoped>
-.alpha {
-  position: fixed;
-  bottom: 0px;
-  left: 0px;
-  width: 150px;
-  height: 50px;
-  background-color: red;
-  color: white;
-}
-</style>
